clean up, ci

+++ src/main/java/com/xkball/x3dmap/client/map/minimap/MinimapHudRenderer.java
@@ -10,18 +10,18 @@ import net.minecraft.client.gui.navigation.ScreenRectangle;
 import org.joml.Vector3f;
 
 public final class MinimapHudRenderer {
-
+    
     private static final WorldMapExtensionServiceImpl HUD_SERVICE = new WorldMapExtensionServiceImpl("");
-
+    
     private MinimapHudRenderer() {
     }
-
+    
     public static void render(GuiGraphicsExtractor graphics, DeltaTracker deltaTracker) {
         var mc = Minecraft.getInstance();
-        if(mc.level == null || mc.player == null || mc.options.hideGui || mc.screen != null) return;
-        if(!ClientConfig.MINIMAP_ENABLED.get()) return;
+        if (mc.level == null || mc.player == null || mc.options.hideGui || mc.screen != null) return;
+        if (!ClientConfig.MINIMAP_ENABLED.get()) return;
         var minimap = MinimapExtension.INSTANCE;
-        if(minimap == null) return;
+        if (minimap == null) return;
         var window = mc.getWindow();
         var guiScaledHeight = window.getGuiScaledHeight();
         var guiScaledWidth = window.getGuiScaledWidth();
@@ -62,7 +62,8 @@ public final class MinimapHudRenderer {
         graphics.submitPictureInPictureRenderState(state);
         MinimapRenderHelper.drawBorder(graphics, x0, y0, x1, y1);
         var b3dGraphics = MinimapRenderHelper.getOrCreateB3dGuiGraphics(graphics);
-        if(HUD_SERVICE.getBooleanState("compass", true)) CompassRenderer.render(b3dGraphics, x0, y0, x1, y1, yRot, 0, 8f);
+        if (HUD_SERVICE.getBooleanState("compass", true))
+            CompassRenderer.render(b3dGraphics, x0, y0, x1, y1, yRot, 0, 8f);
         MinimapPlayerMarker.render(b3dGraphics, x0, y0, x1, y1, player.getYRot(), minimap.rotateWithPlayer());
         var coords = "%d %d %d".formatted(blockPos.getX(), blockPos.getY(), blockPos.getZ());
         var textX = (x0 + x1) / 2f;
@@ -70,7 +71,7 @@ public final class MinimapHudRenderer {
         var font = Minecraft.getInstance().font;
         var cw = font.width(coords);
         var ch = font.lineHeight;
-        graphics.fill((int) (textX - cw / 2f - 1), (int) (textY - 1), (int) (textX + cw / 2f + 1), (int) (textY + ch + 1), 0x90505050);
+        graphics.fill((int) (textX - cw / 2f - 1), textY - 1, (int) (textX + cw / 2f + 1), textY + ch + 1, 0x90505050);
         b3dGraphics.drawCenteredString(coords, textX, textY, 0xCCFFFFFF);
     }
 }
